% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceSelection.R
\name{calcMarkerEff}
\alias{calcMarkerEff}
\title{Calculate marker effects}
\usage{
calcMarkerEff(
  Phenotype,
  Genotype,
  type.geno = "phased",
  Model = "Ridge",
  Effect = "A",
  ...
)
}
\arguments{
\item{Phenotype}{a}

\item{Genotype}{b}

\item{type.geno}{description}

\item{Model}{description}

\item{Effect}{description}

\item{...}{description}
}
\value{
A data frame of pedigree information, sorted in ancestral order.
}
\description{
\code{calcMarkerEff}  Calculate marker effects from either "Ridge", "Lasso", "BRR", "BayesB", or "BayseC". Useder can also choose the effects of marker to be included (additive only or additive + dominance ).
}
