% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceSelection.R
\name{calcSI}
\alias{calcSI}
\title{Calculate selection intensity}
\usage{
calcSI(Marker, Pedigree, genoPhased, nCore = NULL)
}
\arguments{
\item{Marker}{A data frame with 5 columns; MarkerID, Chr, Map, MarkerEff.a, and MarkerEff.d. MarkerID specifies the ID of each marker, and Chr and Map are corresponding to the chromosome number and map position, respectively. MarkerEff.a and MarkerEff.d are the effect of each markers calculated by genomic prediction model. When not considering dominance effect, please fill MarkerEff.d column with 0.}

\item{Pedigree}{A data frame of pedigree information, comprised of three columns: IID, Seed, Pollen. IID column corresponds the individual name, and Seed and Pollen columns correspond to the mother and father name respectively. The idividuals with unknown parents are allowed, and unknown should be assigned a clear name (e.g. NA, "0", "unknown).}

\item{genoPhased}{A matrix of phased genotypic data. Row and columns are corresponding to individuals and markers, respectively. Rownames must be a subset of IID in Pedigree, and Colnames must be a subset of MarkerID in Marker. Each cell must contain two numbers separated by a vertical bar (e.g. "0|1"). In this package, the left number are regarded as mathernal allele, and the right as paternal allele.}

\item{nCore}{Integer indicating the number of cores used for parallel calculation. If not specified, the maximum number of cores minus two is used.}
}
\value{
A list of length three. First element of the list contains the result attributed to the additive effect, and second and third are the result attributed to the dominance and total effects, respectively. When considering only additive effect, length of the list become one. The result in each element are summarized in data frame and has following eight columns;
\describe{
\item{gEffect_s}{Genetic effect of seed parent}
\item{gEffect_p}{Genetic effect of pollen parent}
\item{gEffect_mid}{Genetic effect of mid-parent}
\item{gEffect_o}{Genetic effect of target individual}
\item{Vg_s}{Variance of expected progeny distribution, attributed to the seed parent}
\item{Vg_p}{Variance of expected progeny distribution, attributed to the pollen parent}
\item{Vg}{Variance of expected progeny distribution}
\item{Select_Intens}{Estimated selection intensity of target individual}
}
}
\description{
\code{calcSI}  Calculate selection intensity for all individuals
}
