% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceSelection.R
\name{selectPed}
\alias{selectPed}
\title{Extract pedigree information}
\usage{
selectPed(ped, nameList, missingCode = NA)
}
\arguments{
\item{ped}{A data frame of pedigree information, comprised of three columns: IID, Seed, Pollen. IID column corresponds the individual name, and Seed and Pollen columns correspond to the mother and father name respectively. The idividuals with unknown parents are allowed, and unknown should be assigned a clear name (e.g. NA, "0", "unknown).}

\item{nameList}{A vector of the target individuals whose pedigree information are to be extracted.}

\item{missingCode}{A character assigned for unknown parent. Default value is NA.}
}
\value{
A data frame of pedigree information for specific individuals
}
\description{
\code{selectPed} Extract pedigree information of specific individuals.
}
